changed UnixTime unmarshaler from taking a string to taking a float. modified OHLCResp last field to be uint64 instead of float

+++ bruit/clients/kraken/types/rest_responses.go
@@ -58,7 +58,7 @@ type FeesInfo struct {
 
 type OHLCResp struct {
 	Pair map[string][]RestCandles
-	Last float64 `json:"last"`
+	Last uint64 `json:"last"`
 }
 
 type RestCandles struct {
